requirements.txt
requests>=2.25.1
openai>=1.0.0
asyncio-mqtt>=0.11.1
dataclasses-json>=0.5.7
python-dotenv>=0.19.0
pathlib2>=2.3.6
typing-extensions>=4.0.0
pydantic>=1.8.0
beautifulsoup4>=4.9.3
lxml>=4.6.3
tqdm>=4.61.0
selenium>=4.0.0